Refactoring files, locations, and code
