+++ extension/side-panel.html
@@ -21,10 +21,6 @@
             <!-- Content will be loaded from mainView template -->
         </div>
     </div>
-    <script src="constants.js"></script>
-    <script src="templates.js"></script>
-    <script src="markdown.js"></script>
-    <script src="app.js"></script>
-    <script src="side-panel-init.js"></script>
+    <script src="app.js" type="module"></script>
 </body>
 </html>
