+++ extension/state.js
@@ -1,0 +1,4 @@
+export const state = {
+    conversationHistory: [],
+    contextSize: 4096
+};
